made geo.pl_geometry an int

+++ source/python.h
@@ -454,7 +454,7 @@ int wind_type;		/*Basic prescription for wind(0=SV,1=speherical , 2 can imply ol
   double lum_tot, lum_star, lum_disk, lum_bl, lum_wind;	/* The total luminosities of the disk, star, bl, & wind 
 							   are actually not used in a fundamental way in the program */
   double lum_agn;		/*The total luminosity of the AGN or point source at the center */
-  double pl_geometry;      /* geometry of X-ray point source */
+  int pl_geometry;      /* geometry of X-ray point source */
 #define PL_GEOMETRY_SPHERE 0
 #define PL_GEOMETRY_LAMP_POST 1
   double lamp_post_height; /* height of X-ray point source if lamp post */
